First release of Super Deliveries!
Super Deliveries is a program using the A* search algorithm, the dynamic knapsack solution and custom logic to find the shortest possible route for the highest valuable combination of orders in New York City, instantly. Minimum CPU load, minimum distance traveled.

Orders and addresses are loaded from a SQLite database.
Remember to move the database to a new folder "db" within CWD of main.py!